update typings for Ember.computed macros
